requirements.txt
picamera2
opencv-python
pantilthat
py-telegram-bot-api
python-dotenv